allow a real one process and really traverse the search tree in a depth-first manner.

+++ new_hopy.py
@@ -6,9 +6,10 @@ from random import choice, sample, seed
 from reader import graphs_read
 from time import clock
 from multiprocessing import Process, Queue
+import sys
 
-MULTI = True
-seed = lambda x: None
+MULTI = '-s' not in sys.argv
+# seed = lambda x: None
 
 
 def remove_one_color(motif, color):
@@ -27,17 +28,23 @@ def find_k_path(msg, g, remaining=None, path=None):
 
     def find_sub_path(nodes, initial_guess):
         num = count()
-        Q = [(0, next(num), branch(remaining_motif(initial_guess),
-                                   nodes, initial_guess))]
         k = len(g['motif'])
+        Q = [(k+1-len(initial_guess), next(num),
+              branch(remaining_motif(initial_guess), nodes, initial_guess))]
+        # print "push\t{} ({})".format(k+1, 0)
         solution = []
-        while Q:
-            _, _, r = heappop(Q)
-            for b, u, p, s in r:
-                if b == k:
+        done = False
+        while Q and not done:
+            l, _, r = heappop(Q)
+            # print "POP\t{} ({})".format(l, -l+k+1)
+            for l, u, p, _ in r:
+                if len(p) == k:
                     solution = p
+                    done = True
+                    # print "done"
                     break
-                heappush(Q, (b, next(num), u))
+                heappush(Q, (l, next(num), u))
+                # print "push\t{} ({})".format(l, -l+k+1)
 
         if len(solution) > 0 and not solution[0] < solution[-1]:
             solution.reverse()
@@ -106,7 +113,7 @@ def find_k_path(msg, g, remaining=None, path=None):
         for m, n, p in alt:
             # it seems logical that the color filtering is more efficient when
             # done in next_neighbors because it occurs less often
-            yield len(p), branch(m, n, p), p, n
+            yield k+1-len(p), branch(m, n, p), p, n
 
     seed(13572)
     simplify()
@@ -120,7 +127,10 @@ def find_k_path(msg, g, remaining=None, path=None):
     else:
         remaining_nodes = remaining if remaining is not None else g["nodes"]
         initial_path = path if path is not None else []
-        msg.put((find_sub_path(remaining_nodes, initial_path)))
+        if MULTI:
+            msg.put((find_sub_path(remaining_nodes, initial_path)))
+        else:
+            return find_sub_path(remaining_nodes, initial_path)
 
 if __name__ == '__main__':
     all_inputs = ['example-input', 'no-16-6-1x6', 'no-16-6-2x3', 'no-16-7-1x7',
@@ -129,7 +139,7 @@ if __name__ == '__main__':
                   'unique-16-7-1x7',
                   'complete8', 'tenstars', 'small-no']
 
-    # all_inputs = [all_inputs[0]]
+    # all_inputs = [all_inputs[-3]]
     for testcase in all_inputs:
         with open(testcase+'.txt') as f:
             raw = f.readlines()
@@ -140,29 +150,31 @@ if __name__ == '__main__':
                 seed(13572)
                 t0 = clock()
 
-                msg = Queue()
-                num_answer = 0
-                exist = False
-                path = []
-                pool = []
-                tasks = find_k_path(None, graph) if MULTI else []
-                if tasks == []:
-                    tasks = [(graph['nodes'], [])]
+                if MULTI:
+                    msg = Queue()
+                    num_answer = 0
+                    exist, path = False, []
+                    pool = []
+                    tasks = find_k_path(None, graph) if MULTI else []
+                    if tasks == []:
+                        tasks = [(graph['nodes'], [])]
 
-                for nodes, path in tasks:
-                    pool.append(Process(target=find_k_path,
-                                        args=(msg, graph, nodes, path)))
-                for p in pool:
-                    p.start()
+                    for nodes, path in tasks:
+                        pool.append(Process(target=find_k_path,
+                                            args=(msg, graph, nodes, path)))
+                    for p in pool:
+                        p.start()
 
-                while True:
-                    exist, path = msg.get()
-                    num_answer += 1
-                    if exist or num_answer == len(pool):
-                        break
+                    while True:
+                        exist, path = msg.get()
+                        num_answer += 1
+                        if exist or num_answer == len(pool):
+                            break
 
-                for p in pool:
-                    p.terminate()
+                    for p in pool:
+                        p.terminate()
+                else:
+                    exist, path = find_k_path(0, graph)
 
                 print "{}:{:.3f}s".format(i, clock() - t0)
             print "{}{}".format("yes " if exist else "no",
